Rename topojon.js to addFeature, fix #168
