+++ lib/map-tools/addFeature.js
