+++ lib/map-tools/index.js
@@ -11,8 +11,8 @@ module.exports = function (global) {
     var that = this;
 
     this.addMarker = require('map-tools/addMarker')(global, that);
-    this.addTopoJson = require('map-tools/topojson')(global, that).addTopoJson;
-    this.addGeoJson = require('map-tools/topojson')(global, that).addGeoJson;
+    this.addTopoJson = require('map-tools/addFeature')(global, that).addTopoJson;
+    this.addGeoJson = require('map-tools/addFeature')(global, that).addGeoJson;
     this.updateFeature = require('map-tools/updateFeature')(global, that);
     this.addControl = require('map-tools/addControl')(global, that);
     this.updateMarker = require('map-tools/updateMarker')(global, that);
